Test 65.5%

+++ app/Console/Commands/DeleteGamesAncient.php
@@ -6,10 +6,8 @@ use Illuminate\Console\Command;
 
 class DeleteGamesAncient extends Command
 {
-
-    protected $signature = 'games:delet-ancient';
-
-    protected $description = 'Eliminar juegos cuya fecha de lanzamiento sea anterior al 1 de enero de 1990.';
+    protected $signature = 'games:delete-ancient';
+    protected $description = 'Eliminar juegos cuya fecha de lanzamiento sea anterior a o posterior a una fecha especificada.';
 
     public function __construct()
     {
@@ -18,8 +16,26 @@ class DeleteGamesAncient extends Command
 
     public function handle()
     {
+        $fechaLimite = $this->ask('¿Cuál es la fecha límite para eliminar los juegos? (Formato: yyyy-mm-dd)');
+
+        if (!$this->validateDate($fechaLimite)) {
+            $this->error('La fecha ingresada no tiene el formato correcto (yyyy-mm-dd).');
+            return 1;
+        }
+
+        $orden = $this->choice('¿En qué orden quieres eliminar los juegos?', ['Ascendente', 'Descendente'], 0);
+
+        $orden = ($orden == 'Ascendente') ? 'asc' : 'desc';
+
+        if ($orden == 'asc') {
+            $operador = '>=';
+        } else {
+            $operador = '<=';
+        }
+
         $videojuegos = Videojuego::with(['multimedia', 'generos', 'plataformas', 'precios'])
-            ->where('fecha_lanzamiento', '<=', '1990-01-01')
+            ->where('fecha_lanzamiento', $operador, $fechaLimite)
+            ->orderBy('fecha_lanzamiento', $orden)
             ->get();
 
         $cantidad = $videojuegos->count();
@@ -29,17 +45,26 @@ class DeleteGamesAncient extends Command
             return 0;
         }
 
+        if (!$this->confirm("¿Estás seguro que deseas eliminar {$cantidad} juegos?", true)) {
+            $this->info('Operación cancelada.');
+            return 0;
+        }
+
         foreach ($videojuegos as $videojuego) {
             $videojuego->generos()->detach();
             $videojuego->plataformas()->detach();
-
             $videojuego->precios()->delete();
             $videojuego->multimedia()->delete();
-
             $videojuego->delete();
         }
-        $this->info("Se han eliminado {$cantidad} juegos de más de 1/1/1990.");
+
+        $this->info("Se han eliminado {$cantidad} juegos.");
 
         return 0;
+    }
+    private function validateDate($date)
+    {
+        $d = \DateTime::createFromFormat('Y-m-d', $date);
+        return $d && $d->format('Y-m-d') === $date;
     }
 }
